+++ TShockAPI/Bouncer.cs
@@ -37,13 +37,21 @@ namespace TShockAPI
 	internal sealed class Bouncer
 	{
 		internal Handlers.SendTileSquareHandler STSHandler { get; set; }
+		internal Handlers.NetModules.NetModulePacketHandler NetModuleHandler { get; set; }
+    internal Handlers.EmojiHandler EmojiHandler { get; set; }
 
 		/// <summary>Constructor call initializes Bouncer and related functionality.</summary>
 		/// <returns>A new Bouncer.</returns>
 		internal Bouncer()
 		{
 			STSHandler = new Handlers.SendTileSquareHandler();
-			GetDataHandlers.SendTileSquare += STSHandler.OnReceiveSendTileSquare;
+			GetDataHandlers.SendTileSquare += STSHandler.OnReceive;
+
+			NetModuleHandler = new Handlers.NetModules.NetModulePacketHandler();
+			GetDataHandlers.ReadNetModule += NetModuleHandler.OnReceive;
+
+			EmojiHandler = new Handlers.EmojiHandler();
+			GetDataHandlers.Emoji += EmojiHandler.OnReceiveEmoji;
 
 			// Setup hooks
 			GetDataHandlers.GetSection += OnGetSection;